Apply a given function to specified columns in a data table

Apply a given function to specified columns in a data table

conv_(df, cols, f, ..., newcol)

Arguments

df

a data table

cols

a character vector of column names to which the function f is applied

f

a function that applies to each column.

...

arguments to pass to function f

newcol

a character vector of new column names to which the output of f will be applied.

Details

This function modifies by reference. If passed a column name that does not exist in the provided data table, this will simply return the original data.table unmodified without error.

Examples

x <- data.table::data.table(x = rep("APPLE", 10), y = rep("BANANA", 10), z = rep("CHERRY", 10))
nsqipr:::conv_(x, c("x","y"), tolower)
x
#>         x      y      z
#>  1: apple banana CHERRY
#>  2: apple banana CHERRY
#>  3: apple banana CHERRY
#>  4: apple banana CHERRY
#>  5: apple banana CHERRY
#>  6: apple banana CHERRY
#>  7: apple banana CHERRY
#>  8: apple banana CHERRY
#>  9: apple banana CHERRY
#> 10: apple banana CHERRY

x <- data.table::data.table(x = rep("APPLE", 10), y = rep("BANANA", 10), z = rep("CHERRY", 10))
nsqipr:::conv_(x, c("x","y"), tolower, newcol =  c("X","Y"))
x
#>         x      y      z     X      Y
#>  1: APPLE BANANA CHERRY apple banana
#>  2: APPLE BANANA CHERRY apple banana
#>  3: APPLE BANANA CHERRY apple banana
#>  4: APPLE BANANA CHERRY apple banana
#>  5: APPLE BANANA CHERRY apple banana
#>  6: APPLE BANANA CHERRY apple banana
#>  7: APPLE BANANA CHERRY apple banana
#>  8: APPLE BANANA CHERRY apple banana
#>  9: APPLE BANANA CHERRY apple banana
#> 10: APPLE BANANA CHERRY apple banana

x <- data.table::data.table(x = rep("APPLE", 10), y = rep("BANANA", 10), z = rep("CHERRY", 10))
nsqipr:::conv_(x, "x", paste, "JUICE", sep = " ", newcol = "drink")