Parse entries that indicate the presence of an anastomotic leak

Parse entries that indicate the presence of an anastomotic leak

conv_col_anastomotic(vec)

Arguments

vec

a character vector to parse

Value

a logical vector

Details

returns TRUE if case-insensitive "yes" or "leak' is detected in the character vector.

Examples

x <- c("No", "Yes-reoperation", "Yes-percutaneous intervention",
"Yes-no intervention required",
"No definitive diagnosis of leak/leak related abscess",
"Leak, treated w/ reoperation", "Leak, treated w/ interventional means",
"Leak, no treatment intervention documented",
"Leak, treated w/ non-interventional/non-operative means", NA)

cbind(x, nsqipr:::conv_col_anastomotic(x))
#>       x                                                                
#>  [1,] "No"                                                      "FALSE"
#>  [2,] "Yes-reoperation"                                         "TRUE" 
#>  [3,] "Yes-percutaneous intervention"                           "TRUE" 
#>  [4,] "Yes-no intervention required"                            "TRUE" 
#>  [5,] "No definitive diagnosis of leak/leak related abscess"    "FALSE"
#>  [6,] "Leak, treated w/ reoperation"                            "TRUE" 
#>  [7,] "Leak, treated w/ interventional means"                   "TRUE" 
#>  [8,] "Leak, no treatment intervention documented"              "TRUE" 
#>  [9,] "Leak, treated w/ non-interventional/non-operative means" "TRUE" 
#> [10,] NA                                                        NA