Parse a column for type of anastomotic leak intervention

Parse a column for type of anastomotic leak intervention

conv_col_leak_treatment(vec)

Arguments

vec
a character vector of values to convert

Value

a factor vector

Details

NSQIP encodes the col_anastomotic column as having one of multiple interventions. This function extracts those values from character vectors and factors them.

Examples

x <- c("No", "Yes-reoperation", "Yes-percutaneous intervention",
"Yes-no intervention required",
"No definitive diagnosis of leak/leak related abscess",
"Leak, treated w/ reoperation", "Leak, treated w/ interventional means",
"Leak, no treatment intervention documented",
"Leak, treated w/ non-interventional/non-operative means", NA)

nsqipr:::conv_col_leak_treatment(x)

#>  [1] <NA>                       Reoperation               
#>  [3] Percutaneous intervention  No intervention           
#>  [5] <NA>                       Reoperation               
#>  [7] Percutaneous intervention  No intervention           
#>  [9] Non-operative intervention <NA>                      
#> 4 Levels: Reoperation Percutaneous intervention ... Non-operative intervention