Convert amylase columns from wide to long format

Convert amylase columns from wide to long format

make_amylase_long(df)

Arguments

df

a data.table

Value

a data.table

Details

The data contained in the 3 amylase columns of the targeted pancreatectomy dataset ("pan_amylase_pod1", "pan_amylase_pod230", and "damylase") are converted into a long format. This reduces the total number of columns from 3 to 2 and makes the data more intuitive.

If both "pan_amylase_pod1" and "pan_amylase_pod230" are columns in df, it will be broken into a long format with caseid as the ID variable for joining back to the main table. The amylase values from both "pan_amylase_pod1" and "pan_amylase_pod230" are placed in the "amylase" column". The POD on which this value was acquired (either 1 if from "pan_amylase_pod1" column or set equal to the value of the "damylase" column) is placed in the "pod" column.

Note that this does not alter the original columns.

Examples

x <- data.table::data.table(caseid = 1:30,
pan_amylase_pod1 = c(NA, sample(1:10000, 29)),
pan_amylase_pod230 = c(NA, sample(1000:10000, 29)),
damylase = c(NA, sample(2:30, 29)))

nsqipr:::make_amylase_long(x)
#>     caseid pod amylase
#>  1:      2   1    2797
#>  2:      2   9    1640
#>  3:      3   1    3349
#>  4:      3  12    7253
#>  5:      4   1    8483
#>  6:      4  22    4594
#>  7:      5   1    7948
#>  8:      5  16    3728
#>  9:      6   1    9270
#> 10:      6  10    6474
#> 11:      7   1    9411
#> 12:      7  28    3459
#> 13:      8   1    4725
#> 14:      8   2    9582
#> 15:      9   1    2764
#> 16:      9  23    3813
#> 17:     10   1    8837
#> 18:     10  18    6734
#> 19:     11   1    5398
#> 20:     11   8    3795
#> 21:     12   1    4323
#> 22:     12  26    4143
#> 23:     13   1    3591
#> 24:     13  11    9905
#> 25:     14   1    6647
#> 26:     14  29    6072
#> 27:     15   1    1369
#> 28:     15   6    2705
#> 29:     16   1    6331
#> 30:     16   3    6522
#> 31:     17   1     121
#> 32:     17  25    3651
#> 33:     18   1    6717
#> 34:     18  21    9819
#> 35:     19   1    6315
#> 36:     19  13    9576
#> 37:     20   1    9467
#> 38:     20  27    9230
#> 39:     21   1    3405
#> 40:     21  15    3049
#> 41:     22   1    7246
#> 42:     22  24    6659
#> 43:     23   1    4534
#> 44:     23   5    7401
#> 45:     24   1    9976
#> 46:     24  30    4185
#> 47:     25   1    7336
#> 48:     25  20    3612
#> 49:     26   1    4133
#> 50:     26  14    3992
#> 51:     27   1    5311
#> 52:     27  19    2233
#> 53:     28   1    3312
#> 54:     28  17    1491
#> 55:     29   1     238
#> 56:     29   4    2839
#> 57:     30   1    6413
#> 58:     30   7    6003
#>     caseid pod amylase