Convert pan_percdrainage column from wide to long format

Convert pan_percdrainage column from wide to long format

make_pan_percdrainage_long(df)

Arguments

df
a data.table

Value

a data.table

Details

The data contained in the 4 "pan_percdrainage" columns created by make_pan_percdrainage_cols are converted into a long format.

If "pan_percdrainage" is a column in df, it will be broken into a long format with caseid as the ID variable for joining back to the main table. This is because the targeted pancreatectomy datasets input multiple values into a single "pan_percdrainage" column. For example, "Yes-other, Yes-bile, Yes-pus" may be an entry in the raw data set. This makes parsing for patients that had purulent percutaneous drainage at any point, for example, very difficult.

Note that this does not alter the "pan_percdrainage" column.

Examples

x <- data.table::data.table(caseid = 1:15,
pan_percdrainage = c("Yes-other", "Yes-bile", "Yes-pus", "Yes-amylase-rich fluid",
"Yes-amylase-rich fluid,Yes-pus", "Yes-bile,Yes-other", "Yes-amylase-rich fluid,Yes-pus,Yes-other",
"Yes-amylase-rich fluid,Yes-bile,Yes-other", "Yes-pus,Yes-bile", "Yes-pus,Yes-other",
"Yes-amylase-rich fluid,Yes-pus,Yes-bile", "Yes-amylase-rich fluid,Yes-other",
"Yes-amylase-rich fluid,Yes-pus,Yes-bile,Yes-other", "Yes-amylase-rich fluid,Yes-bile",
"Yes-pus,Yes-bile,Yes-other"))

nsqipr:::make_pan_percdrainage_cols(x)
nsqipr:::make_pan_percdrainage_long(x)

#>     caseid npercdrainage           percdrainage
#>  1:      1             1              Yes-other
#>  2:      2             1               Yes-bile
#>  3:      3             1                Yes-pus
#>  4:      4             1 Yes-amylase-rich fluid
#>  5:      5             1 Yes-amylase-rich fluid
#>  6:      5             2                Yes-pus
#>  7:      6             1               Yes-bile
#>  8:      6             2              Yes-other
#>  9:      7             1 Yes-amylase-rich fluid
#> 10:      7             2                Yes-pus
#> 11:      7             3              Yes-other
#> 12:      8             1 Yes-amylase-rich fluid
#> 13:      8             2               Yes-bile
#> 14:      8             3              Yes-other
#> 15:      9             1                Yes-pus
#> 16:      9             2               Yes-bile
#> 17:     10             1                Yes-pus
#> 18:     10             2              Yes-other
#> 19:     11             1 Yes-amylase-rich fluid
#> 20:     11             2                Yes-pus
#> 21:     11             3               Yes-bile
#> 22:     12             1 Yes-amylase-rich fluid
#> 23:     12             2              Yes-other
#> 24:     13             1 Yes-amylase-rich fluid
#> 25:     13             2                Yes-pus
#> 26:     13             3               Yes-bile
#> 27:     13             4              Yes-other
#> 28:     14             1 Yes-amylase-rich fluid
#> 29:     14             2               Yes-bile
#> 30:     15             1                Yes-pus
#> 31:     15             2               Yes-bile
#> 32:     15             3              Yes-other
#>     caseid npercdrainage           percdrainage