Create readmission columns for long conversion

Create readmission columns for long conversion

make_readm_cols(df)

Arguments

df
a data table to add the columns to

Details

First checks if the data table contains any of the readmission columns. If so, the rest are created as needed. New columns are set to a value of NA.

Examples

x <- data.table::data.table(readmission1 = TRUE)
nsqipr:::make_readm_cols(x)
x

#>    readmission1 readmission2 readmission3 readmission4 readmission5
#> 1:         TRUE           NA           NA           NA           NA
#>    readmpodays1 readmpodays2 readmpodays3 readmpodays4 readmpodays5
#> 1:           NA           NA           NA           NA           NA
#>    readmrelated1 readmrelated2 readmrelated3 readmrelated4 readmrelated5
#> 1:            NA            NA            NA            NA            NA
#>    readmsuspreason1 readmsuspreason2 readmsuspreason3 readmsuspreason4
#> 1:               NA               NA               NA               NA
#>    readmsuspreason5 readmrelicd91 readmrelicd92 readmrelicd93 readmrelicd94
#> 1:               NA            NA            NA            NA            NA
#>    readmrelicd95 readmrelicd101 readmrelicd102 readmrelicd103 readmrelicd104
#> 1:            NA             NA             NA             NA             NA
#>    readmrelicd105 unplannedreadmission1 unplannedreadmission2
#> 1:             NA                    NA                    NA
#>    unplannedreadmission3 unplannedreadmission4 unplannedreadmission5
#> 1:                    NA                    NA                    NA
#>    readmunrelsusp1 readmunrelsusp2 readmunrelsusp3 readmunrelsusp4
#> 1:              NA              NA              NA              NA
#>    readmunrelsusp5 readmunrelicd91 readmunrelicd92 readmunrelicd93
#> 1:              NA              NA              NA              NA
#>    readmunrelicd94 readmunrelicd95 readmunrelicd101 readmunrelicd102
#> 1:              NA              NA               NA               NA
#>    readmunrelicd103 readmunrelicd104 readmunrelicd105
#> 1:               NA               NA               NA