Convert readmission columns from wide to long format

Convert readmission columns from wide to long format

make_readm_long(df, removeFALSE = FALSE)

Arguments

df

a data.table

removeFALSE

a logical vector indicating whether or not to remove rows with a FALSE value.

Value

a data.table

Details

The data from the data table is melted into a long format with caseid as the ID variable to allow rejoining to the main table. After melting, rows with missing values are omitted to reduce the size of the table. Rows where readmission are false may also be removed with removeFALSE to reduce table size if desired, but note this results in an inability to a clarify a known FALSE ("did not have a readmission") from a missing value ("do not know if there was a readmission").

Examples

x <- data.table::data.table(caseid = c(1,2,3,4),
readmission1 = c(TRUE, TRUE, FALSE, NA),
readmpodays1 = c(10, 7, NA, NA),
readmrelated1 = c(TRUE, FALSE, NA, NA),
readmsuspreason1 = c("Reason", "Reason", NA, NA),
readmrelicd91 = c("111", "222", NA, NA),
readmrelicd101 = c("1111","2222", NA, NA),
unplannedreadmission1 = c(TRUE, FALSE, NA, NA),
readmunrelsusp1 = c("Reason", NA, NA, NA),
readmunrelicd91 = c("111", NA, NA, NA),
readmunrelicd101 = c("1111",NA,NA,NA),
readmission2 = c(TRUE, TRUE, FALSE, NA),
readmpodays2 = c(10, 7, NA, NA),
readmrelated2 = c(TRUE, FALSE, NA, NA),
readmsuspreason2 = c("Reason", "Reason", NA, NA),
readmrelicd92 = c("111", "222", NA, NA),
readmrelicd102 = c("1111","2222", NA, NA),
unplannedreadmission2 = c(TRUE, FALSE, NA, NA),
readmunrelsusp2 = c("Reason", NA, NA, NA),
readmunrelicd92 = c("111", NA, NA, NA),
readmunrelicd102 = c("1111",NA,NA,NA),
readmission3 = c(TRUE, TRUE, FALSE, NA),
readmpodays3 = c(10, 7, NA, NA),
readmrelated3 = c(TRUE, FALSE, NA, NA),
readmsuspreason3 = c("Reason", "Reason", NA, NA),
readmrelicd93 = c("111", "222", NA, NA),
readmrelicd103 = c("1111","2222", NA, NA),
unplannedreadmission3 = c(TRUE, FALSE, NA, NA),
readmunrelsusp3 = c("Reason", NA, NA, NA),
readmunrelicd93 = c("111", NA, NA, NA),
readmunrelicd103 = c("1111",NA,NA,NA),
readmission4 = c(TRUE, TRUE, NA, NA),
readmpodays4 = c(10, 7, NA, NA),
readmrelated4 = c(TRUE, FALSE, NA, NA),
readmsuspreason4 = c("Reason", "Reason", NA, NA),
readmrelicd94 = c("111", "222", NA, NA),
readmrelicd104 = c("1111","2222", NA, NA),
unplannedreadmission4 = c(TRUE, FALSE, NA, NA),
readmunrelsusp4 = c("Reason", NA, NA, NA),
readmunrelicd94 = c("111", NA, NA, NA),
readmunrelicd104 = c("1111",NA,NA,NA),
readmission5 = c(TRUE, TRUE, FALSE, NA),
readmpodays5 = c(10, 7, NA, NA),
readmrelated5 = c(TRUE, FALSE, NA, NA),
readmsuspreason5 = c("Reason", "Reason", NA, NA),
readmrelicd95 = c("111", "222", NA, NA),
readmrelicd105 = c("1111","2222", NA, NA),
unplannedreadmission5 = c(TRUE, FALSE, NA, NA),
readmunrelsusp5 = c("Reason", NA, NA, NA),
readmunrelicd95 = c("111", NA, NA, NA),
readmunrelicd105 = c("1111",NA,NA,NA))

nsqipr:::make_readm_long(x)
#>     caseid nreadmission readmission readmpodays unplannedreadmission
#>  1:      1            1        TRUE          10                 TRUE
#>  2:      1            2        TRUE          10                 TRUE
#>  3:      1            3        TRUE          10                 TRUE
#>  4:      1            4        TRUE          10                 TRUE
#>  5:      1            5        TRUE          10                 TRUE
#>  6:      2            1        TRUE           7                FALSE
#>  7:      2            2        TRUE           7                FALSE
#>  8:      2            3        TRUE           7                FALSE
#>  9:      2            4        TRUE           7                FALSE
#> 10:      2            5        TRUE           7                FALSE
#> 11:      3            1       FALSE          NA                   NA
#> 12:      3            2       FALSE          NA                   NA
#> 13:      3            3       FALSE          NA                   NA
#> 14:      3            4       FALSE          NA                   NA
#>     readmrelated readmsuspreason readmunrelsusp readmrelicd9 readmrelicd10
#>  1:         TRUE          Reason         Reason          111          1111
#>  2:         TRUE          Reason         Reason          111          1111
#>  3:         TRUE          Reason         Reason          111          1111
#>  4:         TRUE          Reason         Reason          111          1111
#>  5:         TRUE          Reason         Reason          111          1111
#>  6:        FALSE          Reason           <NA>          222          2222
#>  7:        FALSE          Reason           <NA>          222          2222
#>  8:        FALSE          Reason           <NA>          222          2222
#>  9:        FALSE          Reason           <NA>          222          2222
#> 10:        FALSE          Reason           <NA>          222          2222
#> 11:           NA            <NA>           <NA>         <NA>          <NA>
#> 12:           NA            <NA>           <NA>         <NA>          <NA>
#> 13:           NA            <NA>           <NA>         <NA>          <NA>
#> 14:           NA            <NA>           <NA>         <NA>          <NA>
#>     readmunrelicd9 readmunrelicd10
#>  1:            111            1111
#>  2:            111            1111
#>  3:            111            1111
#>  4:            111            1111
#>  5:            111            1111
#>  6:           <NA>            <NA>
#>  7:           <NA>            <NA>
#>  8:           <NA>            <NA>
#>  9:           <NA>            <NA>
#> 10:           <NA>            <NA>
#> 11:           <NA>            <NA>
#> 12:           <NA>            <NA>
#> 13:           <NA>            <NA>
#> 14:           <NA>            <NA>
nsqipr:::make_readm_long(x, TRUE)
#>     caseid nreadmission readmission readmpodays unplannedreadmission
#>  1:      1            1        TRUE          10                 TRUE
#>  2:      1            2        TRUE          10                 TRUE
#>  3:      1            3        TRUE          10                 TRUE
#>  4:      1            4        TRUE          10                 TRUE
#>  5:      1            5        TRUE          10                 TRUE
#>  6:      2            1        TRUE           7                FALSE
#>  7:      2            2        TRUE           7                FALSE
#>  8:      2            3        TRUE           7                FALSE
#>  9:      2            4        TRUE           7                FALSE
#> 10:      2            5        TRUE           7                FALSE
#>     readmrelated readmsuspreason readmunrelsusp readmrelicd9 readmrelicd10
#>  1:         TRUE          Reason         Reason          111          1111
#>  2:         TRUE          Reason         Reason          111          1111
#>  3:         TRUE          Reason         Reason          111          1111
#>  4:         TRUE          Reason         Reason          111          1111
#>  5:         TRUE          Reason         Reason          111          1111
#>  6:        FALSE          Reason           <NA>          222          2222
#>  7:        FALSE          Reason           <NA>          222          2222
#>  8:        FALSE          Reason           <NA>          222          2222
#>  9:        FALSE          Reason           <NA>          222          2222
#> 10:        FALSE          Reason           <NA>          222          2222
#>     readmunrelicd9 readmunrelicd10
#>  1:            111            1111
#>  2:            111            1111
#>  3:            111            1111
#>  4:            111            1111
#>  5:            111            1111
#>  6:           <NA>            <NA>
#>  7:           <NA>            <NA>
#>  8:           <NA>            <NA>
#>  9:           <NA>            <NA>
#> 10:           <NA>            <NA>