Convert amylase columns from wide to long format

Convert amylase columns from wide to long format

make_amylase_long(df)

Arguments

df
a data.table

Value

a data.table

Details

The data contained in the 3 amylase columns of the targeted pancreatectomy dataset ("pan_amylase_pod1", "pan_amylase_pod230", and "damylase") are converted into a long format. This reduces the total number of columns from 3 to 2 and makes the data more intuitive.

If both "pan_amylase_pod1" and "pan_amylase_pod230" are columns in df, it will be broken into a long format with caseid as the ID variable for joining back to the main table. The amylase values from both "pan_amylase_pod1" and "pan_amylase_pod230" are placed in the "amylase" column". The POD on which this value was acquired (either 1 if from "pan_amylase_pod1" column or set equal to the value of the "damylase" column) is placed in the "pod" column.

Note that this does not alter the original columns.

Examples

x <- data.table::data.table(caseid = 1:30,
pan_amylase_pod1 = c(NA, sample(1:10000, 29)),
pan_amylase_pod230 = c(NA, sample(1000:10000, 29)),
damylase = c(NA, sample(2:30, 29)))

nsqipr:::make_amylase_long(x)

#>     caseid pod amylase
#>  1:      2   1    5293
#>  2:      2  20    2560
#>  3:      3   1    5527
#>  4:      3  13    5590
#>  5:      4   1    6604
#>  6:      4  18    6617
#>  7:      5   1     485
#>  8:      5   6    7308
#>  9:      6   1    2607
#> 10:      6  28    3783
#> 11:      7   1    1476
#> 12:      7  14    3923
#> 13:      8   1    8166
#> 14:      8   7    2874
#> 15:      9   1    2245
#> 16:      9   2    8963
#> 17:     10   1    4613
#> 18:     10  23    2839
#> 19:     11   1    9983
#> 20:     11  25    2314
#> 21:     12   1    4173
#> 22:     12  19    9321
#> 23:     13   1    9090
#> 24:     13  27    4926
#> 25:     14   1    2615
#> 26:     14  12    8017
#> 27:     15   1    3372
#> 28:     15  22    9834
#> 29:     16   1    1981
#> 30:     16   5    2829
#> 31:     17   1    2047
#> 32:     17   8    7595
#> 33:     18   1    3332
#> 34:     18  15    1745
#> 35:     19   1    8944
#> 36:     19   9    3624
#> 37:     20   1    9286
#> 38:     20   3    4841
#> 39:     21   1    6294
#> 40:     21  29    2216
#> 41:     22   1    1023
#> 42:     22  10    7299
#> 43:     23   1    1290
#> 44:     23   4    4083
#> 45:     24   1    9122
#> 46:     24  17    6557
#> 47:     25   1    4276
#> 48:     25  11    2832
#> 49:     26   1    2070
#> 50:     26  30    4649
#> 51:     27   1    1921
#> 52:     27  26    5197
#> 53:     28   1    4275
#> 54:     28  21    1501
#> 55:     29   1    6830
#> 56:     29  24    1772
#> 57:     30   1    4929
#> 58:     30  16    6606
#>     caseid pod amylase